image_segmentator/requirements.txt
Cython==0.28.4
docopt==0.6.2
IPython[all]==6.4.0
Pillow==5.2.0
h5py==2.8.0
imgaug==0.2.6
keras>=2.0.8
-e git+https://github.com/matterport/Mask_RCNN.git@41e7c596ebb83b05a4154bb0ac7a28e0b9afd017#egg=mask_rcnn
numpy==1.15.0
matplotlib==2.2.2
opencv-python==3.4.2.17
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
tensorflow>=1.3.0
